Disable fetching prs when personal access token is restricted
- will reduce very noisy errors
- needs a follow-up to disable integration completely
