+++ apps/desktop/src/lib/forge/github/githubListingService.ts
@@ -1,6 +1,7 @@
 import { DEFAULT_HEADERS } from './headers';
 import { ghResponseToInstance } from './types';
 import { showError } from '$lib/notifications/toasts';
+import { isErrorlike } from '@gitbutler/ui/utils/typeguards';
 import { writable } from 'svelte/store';
 import type { ProjectMetrics } from '$lib/metrics/projectMetrics';
 import type { RepoInfo } from '$lib/url/gitUrl';
@@ -14,6 +15,7 @@ export class GitHubListingService implements ForgeListingService {
 	});
 
 	private error = writable();
+	private disabled = false;
 
 	constructor(
 		private octokit: Octokit,
@@ -22,7 +24,7 @@ export class GitHubListingService implements ForgeListingService {
 	) {}
 
 	async refresh() {
-		if (!navigator.onLine) {
+		if (!navigator.onLine || this.disabled) {
 			return;
 		}
 		try {
@@ -36,6 +38,16 @@ export class GitHubListingService implements ForgeListingService {
 			this.projectMetrics?.setMetric('pr_count', prs.length);
 			this.prs.set(prs);
 		} catch (err: unknown) {
+			// Suppress error if there is a repo restriction on using personal access tokens. This is a
+			// a bit of a hack, and should be reworked into some persistent state that can disable the
+			// integration for a specific repo and give the user appropriate feedback.
+			if (
+				isErrorlike(err) &&
+				err.message.includes('you appear to have the correct authorization credentials')
+			) {
+				this.disabled = true;
+				return;
+			}
 			this.error.set(err);
 			showError('Failed to fetch PRs', err);
 		}
